requirements.txt
googlesearch-python
langchain_ollama
